# bci/openapi.yaml
openapi: 3.0.3
info:
  title: API de simulação — Capital Decrescente (Consumo) BCI
  version: "1.0.0"
  description: |
    Simulação do prémio do seguro de vida associado a crédito ao consumo com
    capital decrescente, e consulta da tabela de tarifas.

    A app não persiste nada: cada pedido é calculado e devolvido.

    ## Regras de cálculo

    | Item | Regra |
    |---|---|
    | Idade | Idade no aniversário mais próximo, à data da simulação |
    | Taxa (‰) | Tabela por perfil × idade (18–70) × prazo (1–7) |
    | Prémio total | `montante × taxa / 1000 × (1 + agravamento)` |
    | Prémio simples | `prémio total / 1,025` |
    | Sobretaxa | 1,5 % do prémio simples |
    | Selo | 1 % do prémio simples |
    | Comissão BCI | 21 % do prémio simples |

    A taxa da tarifa produz o prémio **total** (encargos já incluídos); o prémio
    simples obtém-se por divisão por 1,025 (= 1 + 1,5 % + 1 %). Confirma-se
    depois que `prémio total = prémio simples + sobretaxa + selo`.

    O agravamento é fixo, configurado em `settings.BCI_AGRAVAMENTO`
    (ex.: `0.10` = 10 %); não é um parâmetro do pedido.

    Fonte das regras e das taxas: `Capital Decrescente Consumo Formulas.xlsx`
    e `Capital Decrescente Tarifa BCI.xlsx`.
servers:
  - url: https://emose-serviops.com/api/bci
    description: Produção
  - url: http://127.0.0.1:8010/api/bci
    description: Desenvolvimento local
security:
  - TokenAuth: []
...
paths:
  /processes/quotation:
    post:
      tags: [Processes]
      operationId: simularProcesso
      summary: Simula o prémio de um empréstimo
      description: |
        Calcula o prémio a partir do montante, do perfil do mutuário, da data de
        nascimento e do prazo. Nada é gravado.

        A idade é calculada à data do pedido, logo a mesma `data_nascimento`
        pode dar valores diferentes em datas diferentes.

        Atenção: a rota **não tem barra final**. `/processes/quotation/`
        devolve 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulacaoRequest"
            examples:
              trabalhador:
                summary: Trabalhador, 1.000.000 MZN a 5 anos
                value:
                  tipo_mutuario: TRABALHADORES
                  montante_emprestimo: 1000000
                  data_nascimento: "1996-01-01"
                  prazo_emprestimo: 5
              acentos_e_minusculas:
                summary: tipo_mutuario é normalizado
                value:
                  tipo_mutuario: Mutuários
                  montante_emprestimo: 500000
                  data_nascimento: "1985-07-20"
                  prazo_emprestimo: 3
      responses:
        "200":
          description: Simulação calculada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulacaoResponse"
              example:
                montante_emprestimo: 1000000.0
                premio_total: 11843.99
                premio_simples: 11555.11
                sobre_taxa: 173.33
                selo: 115.55
                comissao_bci: 2426.57
                taxa: 11.84399
        "400":
          description: |
            Erro de validação (um objecto com uma lista de mensagens por campo),
            ou combinação não segurável (objecto com `detail`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErroValidacao"
                  - $ref: "#/components/schemas/ErroDetalhe"
              examples:
                validacao:
                  summary: Campos inválidos
                  value:
                    tipo_mutuario:
                      - "Valor inválido. Use um de: TRABALHADORES, MUTUARIOS, MILITARES."
                    montante_emprestimo:
                      - "Um número válido é necessário."
                    data_nascimento:
                      - "Formato inválido para data. Use um dos formatos a seguir: YYYY-MM-DD."
                    prazo_emprestimo:
                      - "Certifique-se de que este valor seja inferior ou igual a 7."
                nao_seguravel:
                  summary: Idade × prazo sem taxa na tarifa
                  value:
                    detail: "Não segurável: idade 67 com prazo de 7 anos (TRABALHADORES)."
        "401":
          $ref: "#/components/responses/NaoAutenticado"
  /processes/tarif:
    get:
      tags: [Processes]
      operationId: obterTarifa
      summary: Devolve a tabela de tarifas
      description: |
        Taxas em permilagem (‰) por idade (18–70) e prazo em anos (1–7).

        As combinações não seguráveis vêm a `null` em vez de omitidas, para a
        grelha ficar rectangular e poder ser renderizada como tabela
        directamente.

        Atenção: a rota **não tem barra final**. `/processes/tarif/` devolve 404.
      parameters:
        - name: tipo_mutuario
          in: query
          required: false
          description: |
            Perfil a devolver. Omitido, devolve os três.
            Aceita minúsculas e acentos (`Mutuários` → `MUTUARIOS`).
          schema:
            $ref: "#/components/schemas/TipoMutuario"
          example: TRABALHADORES
      responses:
        "200":
          description: Tabela de tarifas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarifaResponse"
              example:
                tipos_mutuario: [TRABALHADORES]
                idade_min: 18
                idade_max: 70
                prazos: [1, 2, 3, 4, 5, 6, 7]
                tarifas:
                  - tipo_mutuario: TRABALHADORES
                    idade: 18
                    taxas:
                      "1": 3.685402
                      "2": 7.114062
                      "3": 10.259172
                      "4": 13.077362
                      "5": 15.622632
                      "6": 17.933739
                      "7": 20.045566
                  - tipo_mutuario: TRABALHADORES
                    idade: 67
                    taxas:
                      "1": 29.777084
                      "2": 59.806923
                      "3": 93.124529
                      "4": 130.097452
                      "5": 171.124289
                      "6": 216.64395
                      "7": null
        "400":
          description: "`tipo_mutuario` inválido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDetalhe"
              example:
                detail: "tipo_mutuario inválido. Use um de: TRABALHADORES, MUTUARIOS, MILITARES."
        "401":
          $ref: "#/components/responses/NaoAutenticado"
  /processes/simulate-offline/:
    post:
      tags: [Processes]
      operationId: registarVendaOffline
      summary: Regista uma venda fechada ao balcão
      description: |
        **Stub.** Não valida o risco, não calcula prémio e não decide nada:
        guarda o pedido e responde sempre aprovado. Quem emite a apólice é o
        RPA INSIS, a jusante.

        A resposta é constante, tirando `id_bci`, que é devolvido tal como veio
        no pedido. Sem `id_bci` no pedido, vem string vazia.

        Qualquer corpo JSON é aceite, incluindo vazio. Não há campos
        obrigatórios, e valores fora dos limites da tarifa (por exemplo
        `prazo_emprestimo: 12`) não são rejeitados, porque não há cálculo.
        Tudo o que o canal enviar é guardado tal como veio.

        Quando o pedido traz `transaction_ref`, é idempotente: repetir não cria
        um segundo registo. A resposta é a mesma nos dois casos.

        A rota responde **com e sem** barra final.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistoOfflineRequest"
            example:
              id_bci: "EMOSETESTE0063"
              transaction_ref: "TRX-2026-000841"
              tipo_mutuario: "trabalhador"
              montante_emprestimo: 1e+06
              data_nascimento: "1995-01-21"
              prazo_emprestimo: 12
              agravamento: 0
              nuit: "125454958"
              nome_cliente: "Raimundo L. Timba"
              apelido: "Timba"
              tipo_entidade: "person"
              sexo: "M"
              pais: "Moçambique"
              nacionalidade: "MZ"
              lingua: "PORTUGUESE"
              tipo_morada: "Casa"
              provincia: "Maputo Cidade"
              regiao: "Distrito Municipal de KaMpfumo"
              bairro: "Polana Cimento A"
              morada: "Av. Julius Nyerere 123"
              tipo_contacto: "Mobile"
              numero_contacto: "847527898"
              lingua_contacto: "Portugues"
              codigo_banco: "0008"
              numero_conta: "123456789012"
              tipo_moeda: "MZN"
              tipo_seguro: "vida"
              canal_venda: "balcao"
              data_emissao: "2026-01-18"
              data_apolice: "2026-01-18"
              data_inicio_seguro: "2026-01-18"
              papel: "tomador"
              quota: 100
              estado: "activo"
              designacao: "Seguro Vida Credito Consumo"
              lingua_apolice: "Portugues"
              tipo_pagamento: "debito_conta"
              num_prestacoes: 1
              man_id: "6000035452"
              tipo_segurado: "principal"
              capital_seguro: 1e+06
              premio_pago: 11844
              comissao: 2426.57
              certificado_apolice: "JVBERi0xLjQKJeLjz9MK..."
              altura_cm: 175
              peso_kg: 70
              fuma: "Sim"
              fuma_anos_cigarros: "12 anos, 5 por dia"
              bebe_bebidas_alcoolicas: "Sim"
              bebe_frequencia: "Ao fim-de-semana"
              pratica_desporto: "Nao"
              tipo_desporto: ""
              toma_drogas_medicamentos: "Nao"
              tipo_tempo_medicamentos: ""
              af_diabetes: "Sim"
              af_colesterol_alto: "Nao"
              af_doencas_cardiacas: "Nao"
              af_avc: "Sim"
              af_hipertensao: "Nao"
              af_doenca_neurologica_mental: "Nao"
              af_cancro: "Nao"
              af_retinite_pigmentosa: "Nao"
              af_porfiria: "Nao"
              af_hemofilia: "Nao"
              af_outra_doenca_hereditaria: "Nao"
              af_morte_familiar_menos_65: "Nao"
              ap_tratamento_cirurgico: "Sim"
              ap_tratamento_cirurgico_especifique: "Apendicectomia em 2019"
              ap_diagnostico_5anos: "Nao"
              ap_diagnostico_5anos_descreva: ""
              ap_febre_reumatica_coracao: "Nao"
              ap_pressao_arterial_avc: "Nao"
              ap_doenca_pulmonar: "Nao"
              ap_sistema_digestivo: "Nao"
              ap_condicoes_rins: "Nao"
              ap_ansiedade_depressao: "Sim"
              ap_diabetes_glandular: "Nao"
              ap_articulacoes_coluna: "Nao"
              ap_tumores: "Nao"
              ap_doencas_sangue: "Nao"
              ap_hiv: "Nao"
              ap_hiv_indique: ""
              ap_doenca_sexual: "Nao"
              ap_deficiencia: "Nao"
              ap_circunstancias_risco: "Nao"
              ap_outra_doenca: "Nao"
              ap_outra_doenca_mencione: ""
              sf_esta_gravida: "Nao"
              sf_gravida_meses: ""
              sf_esteve_gravida: "Nao"
              sf_esteve_gravida_detalhes: ""
              sf_tipo_partos: ""
              sf_numero_partos_anos: ""
              sf_aborto_prematuro: "Nao"
              sf_aborto_meses: ""
              sf_doenca_ginecologica: "Nao"
              sf_doenca_ginecologica_mencione: ""
      responses:
        "200":
          description: Pedido aceite e guardado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistoOfflineResponse"
              example:
                message: "Cotação aprovada. RPA INSIS iniciado para emissão de apólice."
                id_bci: "EMOSETESTE0063"
                estado_processo: "aprovado"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Token do Django REST Framework (`rest_framework.authtoken`).
        O valor do header inclui o prefixo `Token`:

        ```
        Authorization: Token c183b8795ebcb774bdc2491d28286bdcb3f4bec1
        ```

        Os tokens vivem na tabela `authtoken_token` da base de dados. Um token
        por utilizador. Emitir com
        `python manage.py drf_create_token <username>`.
  responses:
    NaoAutenticado:
      description: Header `Authorization` ausente ou token inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroDetalhe"
          example:
            detail: "As credenciais de autenticação não foram fornecidas."
  schemas:
    TipoMutuario:
      type: string
      description: |
        Perfil do mutuário, que determina a tabela de taxas a usar.
        Normalizado antes de ser validado: maiúsculas e minúsculas são
        equivalentes e os acentos são removidos.
      enum: [TRABALHADORES, MUTUARIOS, MILITARES]
    SimulacaoRequest:
      type: object
      required:
        [tipo_mutuario, montante_emprestimo, data_nascimento, prazo_emprestimo]
      properties:
        tipo_mutuario:
          $ref: "#/components/schemas/TipoMutuario"
        montante_emprestimo:
          type: number
          format: double
          description: Capital do empréstimo. Máximo 16 dígitos, 2 casas decimais.
          minimum: 0.01
          example: 1000000
        data_nascimento:
          type: string
          format: date
          description: |
            Só o formato `YYYY-MM-DD` é aceite. Tem de ser no passado, e a idade
            no aniversário mais próximo tem de cair entre 18 e 70 anos.
          example: "1996-01-01"
        prazo_emprestimo:
          type: integer
          description: Duração do empréstimo em anos.
          minimum: 1
          maximum: 7
          example: 5
    SimulacaoResponse:
      type: object
      required:
        [
          montante_emprestimo,
          premio_total,
          premio_simples,
          sobre_taxa,
          selo,
          comissao_bci,
          taxa,
        ]
      properties:
        montante_emprestimo:
          type: number
          format: double
          description: Eco do capital recebido no pedido.
        premio_total:
          type: number
          format: double
          description: >-
            `montante × taxa / 1000 × (1 + agravamento)`.
            Igual a prémio simples + sobretaxa + selo.
        premio_simples:
          type: number
          format: double
          description: "`prémio total / 1,025`."
        sobre_taxa:
          type: number
          format: double
          description: 1,5 % do prémio simples.
        selo:
          type: number
          format: double
          description: 1 % do prémio simples (imposto de selo).
        comissao_bci:
          type: number
          format: double
          description: 21 % do prémio simples. Não entra no prémio total.
        taxa:
          type: number
          format: double
          description: |
            Taxa da tarifa em permilagem (‰) usada no cálculo, para se poder
            conferir o resultado: `premio_total = montante × taxa / 1000`
            (com agravamento 0).
...
    ErroDetalhe:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ErroValidacao:
      type: object
      description: |
        Uma entrada por campo inválido, com a lista de mensagens.

        As mensagens genéricas do Django REST Framework seguem o
        `LANGUAGE_CODE` do projecto, que em `config/settings.py` é `pt` — os
        exemplos são os dessa configuração. Noutra instalação com outro
        `LANGUAGE_CODE` o texto muda, pelo que os clientes devem decidir pela
        chave do campo e não pela mensagem. As mensagens escritas nesta app
        (perfil inválido, não segurável) estão sempre em português.
      additionalProperties:
        type: array
        items:
          type: string
...